pyproject.toml
[project]
name = "sitcom-simulator"
version = "0.0.12"
authors = [
    { name = "Josh Moody", email = "josh@joshmoody.org" },
]
description = "A highly-customizable tool for automatically creating AI-generated meme videos"
readme = "README.md"
license = {file = "LICENSE"}
keywords = ["ai", "video", "meme", "generator", "sitcom", "simulator"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent", 
]
dependencies = [
    "stability-sdk~=0.8.4",
    "python-dotenv~=1.0.0",
    "tqdm~=4.66.1",
    "openai~=0.28.0",
    "ffmpeg-python~=0.2.0",
    "gtts~=2.3.2",
    "Pillow~=10.0.1",
    "beautifulsoup4~=4.12.2",
    "requests~=2.31.0",
    # "fakeyou==1.2.5", Currently using raw HTTP requests instead
    # "moviepy==1.0.3", No longer supported due to lack of features. Using ffmpeg-python instead
]
requires-python = ">= 3.11"

[project.optional-dependencies]
dev = [
    "mypy",
    "sphinx",
    "furo",
    "sphinx-autodoc-typehints",
    "mypy",
]

[project.urls]
homepage = "https://github.com/joshmoody24/sitcom-simulator"
documentation = "https://joshmoody24.github.io/sitcom-simulator/"
repository = "https://github.com/joshmoody24/sitcom-simulator"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
include = ["sitcom_simulator/**/*"]

[project.scripts]
sitcom-simulator = "sitcom_simulator.cli:main"